# swagger.json
{
  "swagger": "2.0",
    "info": {
        "title": "Smart Waste Segregator",
        "description" : "Smart Waste Segregator API information",
        "contact" : {
        "name" : "SWS Team",
        "email" : "helpsws2.0@gmail.com",
        "url" : "https://smartwastesegregator.netlify.app/"
        },
        "servers": [
        {
            "url" : "http://localhost:7000",
            "description" : "Local Server"
        },
        {
            "url" : "https://helpsws.herokuapp.com/",
            "description" : "Production Server"
        }
        ]
    }
...
  "paths": {
...
    "/updateUserTrack": {
        "post":{
              "summary":"Update user track(recycled/scanned) on the basis of email and type of task",
              "tags": ["UserTrack"],
              "parameters": [
                    {
                    "name": "email",
                    "in": "body",
                    "description": "Email of user"
                    },
                    {
                    "name": "task",
                    "in": "body",
                    "description": "Task to increment(recycled/scanned)"
                    }
                ],
              "responses":{
                  "200":{
                      "description":"OK"
                  }
              }
        }
    },
    "/userTrack": {
        "post":{
              "summary":"Fetch user's track data on the basis of email and type of task",
              "tags": ["UserTrack"],
              "parameters": [
                    {
                    "name": "email",
                    "in": "body",
                    "description": "Email of user"
                    }
                ],
              "responses":{
                  "200":{
                      "description":"OK"
                  }
              }
        }
    }
  }
}